Add TrainingRemoteApiDataSource
